fix develop (StateBoardコンポーネントに分離)
